+++ src/lib.rs
@@ -168,6 +168,8 @@ pub use crate::xdg::{
     NotificationHandle,
     dbus_stack,
     DbusStack,
+    CloseHandler,
+    CloseReason,
 };
 
 #[cfg(all(feature = "server", unix, not(target_os = "macos")))]